light scroll control
instead of moving back to the top of the page every time an overlay is exited, the page is now moved so the top of the image that was just opened is at the top of the page. It's not perfect, but it is good enough and keeps the image the user was just viewing in view when the overlay is closed.